move dendrogram outside of clustering view
