+++ lexos/templates/base.html
@@ -152,8 +152,8 @@
 	              <li><a href="{{ url_for('statistics.statistics') }}">Statistics</a></li>
 	              <li><a href="#">Clustering <span class="caret caret-right"></span></a>
 	                <ul class="dropdown-menu">
-	                  <li><a href="{{ url_for('cluster.dendrogram') }}">Hierarchical Clustering</a></li>
-	                  <li><a href="{{ url_for('cluster.k_means') }}">K-Means Clustering</a></li>
+	                  <li><a href="{{ url_for('dendrogram.dendrogram') }}">Hierarchical Clustering</a></li>
+	                  <li><a href="{{ url_for('k_means.k_means') }}">K-Means Clustering</a></li>
 	                </ul>
 	              </li>
 	              <li><a href="{{ url_for('sim_query.similarity') }}">Similarity Query</a></li>
