+++ lexos/views/k_means_view.py
@@ -1,4 +1,3 @@
-import json
 import os
 from os.path import join as path_join
 
@@ -6,7 +5,6 @@ from flask import send_file, request, session, render_template, Blueprint
 
 from lexos.helpers import constants as constants
 from lexos.managers import session_manager as session_manager, utility
-from lexos.models.dendro_model import DendrogramModel
 from lexos.views.base_view import detect_active_docs
 
 # this is a flask blue print
@@ -14,40 +12,11 @@ from lexos.views.base_view import detect_active_docs
 # see here for more detail:
 # http://exploreflask.com/en/latest/blueprints.html
 # http://flask.pocoo.org/docs/0.12/blueprints/
-cluster_blueprint = Blueprint('cluster', __name__)
-
-
-@cluster_blueprint.route("/dendrogram", methods=['GET'])
-def dendrogram():
-    # Detect the number of active documents.
-    num_active_docs = detect_active_docs()
-    file_manager = utility.load_file_manager()
-    if 'analyoption' not in session:
-        session['analyoption'] = constants.DEFAULT_ANALYZE_OPTIONS
-    if 'hierarchyoption' not in session:
-        session['hierarchyoption'] = constants.DEFAULT_HIERARCHICAL_OPTIONS
-    labels = file_manager.get_active_labels()
-    return render_template(
-        'dendrogram.html',
-        labels=labels,
-        numActiveDocs=num_active_docs,
-        itm="hierarchical")
-
-
-@cluster_blueprint.route("/dendrogramDiv", methods=['POST'])
-def dendrogram_div():
-    """Send the dendrogram div that is generated by plotly to frontend
-
-    :return: an html string of the dendrogram div
-    """
-    session_manager.cache_analysis_option()
-    session_manager.cache_hierarchy_option()
-    return json.dumps(
-        {'dendroDiv': DendrogramModel().get_dendrogram_div()})
+k_means_blueprint = Blueprint('k_means', __name__)
 
 
 # Tells Flask to load this function when someone is at '/kmeans'
-@cluster_blueprint.route("/kmeans", methods=["GET", "POST"])
+@k_means_blueprint.route("/kmeans", methods=["GET", "POST"])
 def k_means():
     """Handles the functionality on the kmeans page.
 
@@ -128,7 +97,7 @@ def k_means():
 
 
 # Tells Flask to load this function when someone is at '/kmeansimage'
-@cluster_blueprint.route("/kmeansimage", methods=["GET", "POST"])
+@k_means_blueprint.route("/kmeansimage", methods=["GET", "POST"])
 def k_means_image():
     """Reads the png image of the kmeans and displays it on the web browser.
 
@@ -145,7 +114,7 @@ def k_means_image():
     return send_file(image_path)
 
 
-@cluster_blueprint.route("/small_PCA", methods=["GET", "POST"])
+@k_means_blueprint.route("/small_PCA", methods=["GET", "POST"])
 def small_pca():
     """Reads the small image of the PCA and displays it on the web browser.
 
@@ -160,7 +129,7 @@ def small_pca():
         return send_file(plotly_url)
 
 
-@cluster_blueprint.route("/big_PCA", methods=["GET", "POST"])
+@k_means_blueprint.route("/big_PCA", methods=["GET", "POST"])
 def big_pca():
     """Reads the big image of the PCA and displays it on the web browser.
